// react-module/src/openapi/models/atomic-product-definition-extended-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface AtomicProductDefinitionExtendedDto
 */
export interface AtomicProductDefinitionExtendedDto {
    /**
     * 
     * @type {number}
     * @memberof AtomicProductDefinitionExtendedDto
     */
    id?: any;
    /**
     * 
     * @type {string}
     * @memberof AtomicProductDefinitionExtendedDto
     */
    title?: any;
    /**
     * 
     * @type {AtomicProductDefinitionExtendedDto}
     * @memberof AtomicProductDefinitionExtendedDto
     */
    parent?: any;
    /**
     * 
     * @type {Array&lt;AtomicProductParameterExtendedDto&gt;}
     * @memberof AtomicProductDefinitionExtendedDto
     */
    parameters?: any;
    /**
     * 
     * @type {GidDto}
     * @memberof AtomicProductDefinitionExtendedDto
     */
    gid?: any;
    /**
     * 
     * @type {LayerShapeExtendedDto}
     * @memberof AtomicProductDefinitionExtendedDto
     */
    layerShape?: any;
    /**
     * 
     * @type {string}
     * @memberof AtomicProductDefinitionExtendedDto
     */
    category?: AtomicProductDefinitionExtendedDtoCategoryEnum;
}

/**
    * @export
    * @enum {string}
    */
export enum AtomicProductDefinitionExtendedDtoCategoryEnum {
    BASE = 'BASE',
    INTERMEDIATE = 'INTERMEDIATE',
    FINAL = 'FINAL'
}