// react-module/src/openapi/models/atomic-product-parameter-basic-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface AtomicProductParameterBasicDto
 */
export interface AtomicProductParameterBasicDto {
    /**
     * 
     * @type {number}
     * @memberof AtomicProductParameterBasicDto
     */
    id?: any;
    /**
     * 
     * @type {string}
     * @memberof AtomicProductParameterBasicDto
     */
    value?: any;
    /**
     * 
     * @type {AtomicProductParameterTypeBasicDto}
     * @memberof AtomicProductParameterBasicDto
     */
    key?: any;
}