// react-module/src/openapi/models/atomic-product-parameter-extended-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface AtomicProductParameterExtendedDto
 */
export interface AtomicProductParameterExtendedDto {
    /**
     * 
     * @type {AtomicProductDefinitionExtendedDto}
     * @memberof AtomicProductParameterExtendedDto
     */
    atomicProductDefinition?: any;
    /**
     * 
     * @type {number}
     * @memberof AtomicProductParameterExtendedDto
     */
    id?: any;
    /**
     * 
     * @type {string}
     * @memberof AtomicProductParameterExtendedDto
     */
    value?: any;
    /**
     * 
     * @type {AtomicProductParameterTypeBasicDto}
     * @memberof AtomicProductParameterExtendedDto
     */
    key?: any;
}