// react-module/src/openapi/models/atomic-product-parameter-type-basic-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface AtomicProductParameterTypeBasicDto
 */
export interface AtomicProductParameterTypeBasicDto {
    /**
     * 
     * @type {number}
     * @memberof AtomicProductParameterTypeBasicDto
     */
    id?: any;
    /**
     * 
     * @type {string}
     * @memberof AtomicProductParameterTypeBasicDto
     */
    title?: any;
}